refactor: use useOrgBoard, refactor AboutEntity
